requirements.txt
pyautogui
mss
pillow
numpy
keyboard
opencv-python
pytesseract